extract get_absolute_path to helper
